+++ packages/cli/test/unit/commands/index.test.ts
@@ -4,8 +4,10 @@ describe('index', () => {
   it('outputs a map of what aliases our commands use', () => {
     expect(commands).toEqual(
       new Map([
+        ['agent', 'agent'],
         ['alias', 'alias'],
         ['aliases', 'alias'],
+        ['activity', 'activity'],
         ['api', 'api'],
         ['bisect', 'bisect'],
         ['blob', 'blob'],